Added compatibility to access entire fishbase/sealifebase database, not just matched cases from ARTIS
Next steps: Add additional fishbase genus / species presences that are of commercial interest into ARTIS to improve capabilities for interpolating taxa.

In this submission, I also made it so that the server doesn't have to be fetched every time we load in the fishbase data - I created a separate R script to write all the fishbase data into .parquet files.