Successfully acquired additional FB/SLB commercial species that pertain to ARTIS countries
This is the good news, and there are 2991 additional species that will get added to ARTIS! The thing to consider is that the fishbase only includes up to genus names for each sciname, and not only these species names have taxonomic information covered in the "sciname" dataset. My idea is to obtain taxonomic data from rgbif for each of these new species, find the most common taxa names at each level, combine with the species that ARTIS, which should certainly improve the robustness of the amount of presences each country has. I am doing taxonomic acquisition using the package "rgbif," and have created a separate script file for this.